update: fetch nginx package more efficiently via alpine version
This commit updates the method of fetching the latest nginx package and its release version. Instead of using two separate wget commands and extracting the package release from the HTML content, it now directly fetches the package list for the specific Alpine version and extracts the package release number from it. This approach simplifies the code and makes it more efficient.